rex: change u64 termination_flag to u8
Fixes: 48e2bbcb4913 ("rex: add termination_check in base_helper")

+++ rex/src/stub.rs
@@ -222,5 +222,5 @@ extern "C" {
     ///  
     ///  Used to indidicate whether a BPF program in a CPU is executing
     ///  inside a helper, or inside a panic handler, or just in BPF text.
-    pub(crate) static mut bpf_termination_flag: u64;
+    pub(crate) static mut bpf_termination_flag: u8;
 }
